some progress w/predict
